+++ src/00_config.cpp
@@ -1,23 +1,16 @@
-#include "cpl_conv.h"
 #include <Rcpp.h>
+#include "gdallibrary/gdallibrary.h"
 using namespace Rcpp;
+using namespace gdallibrary;
 
 // [[Rcpp::export]]
-IntegerVector gdal_set_config_option(CharacterVector option, CharacterVector value) 
+IntegerVector set_gdal_config_cpp(CharacterVector option, CharacterVector value) 
 {
- CPLSetConfigOption( option[0], value[0] );
-  
-
- return 1;
+  gdal_set_config_option( option, value );
+  return 1;
 }
 
 // [[Rcpp::export]]
-CharacterVector gdal_get_config_option(CharacterVector option){
-  CharacterVector out(1);
-  const char *str = CPLGetConfigOption(option[0], nullptr);
-  if (str) 
-    {
-    out[0] = str;
-    }
-  return out;
+CharacterVector get_gdal_config_cpp(CharacterVector option){
+  return gdal_get_config_option(option);
 }